compiler warnings fixes
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/branches/pcl-1.x@4955 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ search/test/test_kdtree.cpp
@@ -51,22 +51,22 @@ PointCloud<PointXYZ> cloud, cloud_big;
 void
 init ()
 {
-  float resolution = 0.1;
+  float resolution = 0.1f;
   for (float z = -0.5f; z <= 0.5f; z += resolution)
     for (float y = -0.5f; y <= 0.5f; y += resolution)
       for (float x = -0.5f; x <= 0.5f; x += resolution)
         cloud.points.push_back (PointXYZ (x, y, z));
-  cloud.width = cloud.points.size ();
+  cloud.width = static_cast<uint32_t> (cloud.points.size ());
   cloud.height = 1;
 
   cloud_big.width = 640;
   cloud_big.height = 480;
-  srand (time (NULL));
+  srand (static_cast<unsigned int> (time (NULL)));
   // Randomly create a new point cloud
   for (size_t i = 0; i < cloud_big.width * cloud_big.height; ++i)
-    cloud_big.points.push_back (
-                                PointXYZ (1024 * rand () / (RAND_MAX + 1.0), 1024 * rand () / (RAND_MAX + 1.0),
-                                          1024 * rand () / (RAND_MAX + 1.0)));
+    cloud_big.points.push_back (PointXYZ (static_cast<float> (1024 * rand () / (RAND_MAX + 1.0)), 
+                                          static_cast<float> (1024 * rand () / (RAND_MAX + 1.0)),
+                                          static_cast<float> (1024 * rand () / (RAND_MAX + 1.0))));
 }
 
 /* Test for KdTree nearestKSearch */TEST (PCL, KdTree_nearestKSearch)
@@ -79,7 +79,7 @@ init ()
   for (size_t i = 0; i < cloud.points.size (); ++i)
   {
     float distance = euclideanDistance (cloud.points[i], test_point);
-    sorted_brute_force_result.insert (make_pair (distance, (int)i));
+    sorted_brute_force_result.insert (make_pair (distance, static_cast<int> (i)));
   }
   float max_dist = 0.0f;
   unsigned int counter = 0;
